♻️ refactor(booksentence): correct spelling mistake in file path and references
- rename "setence" to "sentence" for all related components and paths
