+++ apps/tk-capture-book/app/(auth)/sentence.tsx
@@ -4,8 +4,8 @@ import { useAtomValue } from "jotai";
 import { useTranslation } from "react-i18next";
 
 import { SelectedBookHeader } from "@/features/book/components/selected-book-header";
-import { BookSentence } from "@/features/book/components/setence/book-sentence";
-import { BookSentenceGuide } from "@/features/book/components/setence/book-sentence-guide";
+import { BookSentence } from "@/features/book/components/sentence/book-sentence";
+import { BookSentenceGuide } from "@/features/book/components/sentence/book-sentence-guide";
 import { selectedBookAtom } from "@/features/book/states/book";
 
 export default function SentenceScreen() {
